requirements.txt
requests
PIL
transformers
google-generativeai
gtts
googletrans
streamlit